Modularize Server
